refactor: UserAccountPage 뷰와 비즈니스 로직 분리
