+++ src/pages/AccountManagePage/UserAccountPage/index.tsx
@@ -1,4 +1,3 @@
-import { useState, useEffect } from 'react';
 import { Table, Thead, Tbody, Tr, Td, Th, TableContainer, Box } from '@chakra-ui/react';
 import { SettingsIcon } from '@chakra-ui/icons';
 import { BanState } from '../style';
@@ -6,51 +5,20 @@ import RightContainer from '../../../components/layout/RightContainer';
 import { BanAlertDialogModal } from '../../../components/features/AlertDialogModal';
 import AccountManageModal from '../../../components/features/AccountManageModal';
 import Pagination from '../../../components/features/Pagination';
-import toast from 'react-hot-toast';
-import { useGetUserAccountRequest, useGetTotalAccountCountRequest } from '../../../api-requests/Account';
-import { useSelectedAccountAtom } from '../../../store/accountAtom';
-import { IAccountData } from '../../../@types/Account';
+import useUserAccount from './useUserAccount';
 
 function UserAccountPage() {
-  const [isAlertOpen, setIsAlertOpen] = useState(false);
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
-  const [currentPage, setCurrentPage] = useState(1);
-  const [totalPosts, setTotalPosts] = useState(1);
-  const [accountList, setAccountList] = useState<IAccountData[] | null>([]);
-  const [userCount, setUserCount] = useState(0);
-  const { setSelectedAccount } = useSelectedAccountAtom();
-
   const {
-    data: accountResult,
-    isError: accountError,
-    refetch: accountRefetch,
-  } = useGetUserAccountRequest({ page: currentPage - 1 }, false);
-
-  const { totalAccountCountRefetch } = useGetTotalAccountCountRequest(false);
-
-  useEffect(() => {
-    void accountRefetch();
-  }, [currentPage]);
-
-  useEffect(() => {
-    void totalAccountCountRefetch();
-  }, []);
-
-  useEffect(() => {
-    if (accountResult) {
-      setAccountList(accountResult?.data.memberList);
-      setTotalPosts(accountResult?.data.accountCount);
-    } else if (accountError) {
-      toast.error('유저 계정 목록을 불러오는데 실패했습니다.');
-    }
-  }, [accountResult, accountError]);
-
-  const handleAccountManageModal = (account: IAccountData) => {
-    setSelectedAccount(account);
-    setIsModalOpen(true);
-  };
-
+    isAlertOpen,
+    setIsAlertOpen,
+    isModalOpen,
+    setIsModalOpen,
+    currentPage,
+    setCurrentPage,
+    accountList,
+    totalPosts,
+    handleAccountManageModal,
+  } = useUserAccount();
   return (
     <RightContainer title={'계정 관리'}>
       <BanAlertDialogModal
@@ -66,7 +34,7 @@ function UserAccountPage() {
         setCurrentPage={setCurrentPage}
       />
       <Box mb="10px" fontWeight="600" color="#25549c">
-        유저 계정 관리 (총 계정 수: {userCount})
+        유저 계정 관리 (총 계정 수: {(accountList && accountList.length) || 0})
       </Box>
       <TableContainer>
         <Table variant="simple">
